ref: move SelectCombobox subcomponents into directory
