+++ frontend/src/components/Dropdown/elements/SelectCombobox/index.ts
@@ -1,0 +1,1 @@
+export { SelectCombobox } from './SelectCombobox'
